requirements.txt
Django==4.1.7
djangorestframework==3.14.0
ipython[all]>=3.2.0
pyzmq>=14.7.0
Pillow>=2.9.0
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19